Se incluye el player_id dentro de los comandos enviados para que el server pueda reconocer quien hizo el input

+++ client/client.h
@@ -27,7 +27,7 @@
 class Client {
 private:
     ClientProtocol protocol;
-    int my_player_id;
+    uint8_t my_player_id;
 
     Queue<InputCmd> input_queue;
     Queue<GameStateDTO> state_queue;
