+++ common/clientCommand.h
@@ -5,7 +5,7 @@
 
 //Struct a utilizar para enviar comandos
 struct InputCmd {
-    uint32_t player_id;
+    uint8_t player_id;
     InputKey key;
     InputAction action;
